% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkInstall}
\alias{checkInstall}
\title{Checks the plp package is installed sufficiently for the network study and does other checks if needed}
\usage{
checkInstall(connectionDetails = NULL)
}
\arguments{
\item{connectionDetails}{The connections details for connecting to the CDM}
}
\value{
A number (a value other than 1 means an issue with the install)
}
\description{
Checks the plp package is installed sufficiently for the network study and does other checks if needed
}
\details{
This will check that the network study dependancies work
}
