% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{execute}
\alias{execute}
\title{Execute the validation study}
\usage{
execute(connectionDetails, databaseName, cdmDatabaseSchema,
  cohortDatabaseSchema, oracleTempSchema, cohortTable, outputFolder,
  createCohorts = T, runValidation = T, packageResults = T,
  minCellCount = 5, sampleSize = NULL)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{databaseName}{A string representing a shareable name of your databasd}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{createCohorts}{Whether to create the cohorts for the study}

\item{runValidation}{Whether to run the valdiation models}

\item{packageResults}{Whether to package the results (after removing sensitive details)}

\item{minCellCount}{The min count for the result to be included in the package results}

\item{sampleSize}{Whether to sample from the target cohort - if desired add the number to sample}
}
\description{
Execute the validation study
}
\details{
This function will execute the sepcified parts of the study
}
