% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transportPlpModels}
\alias{transportPlpModels}
\title{Transport trained PLP models into the validation package}
\usage{
transportPlpModels(analysesDir, minCellCount = 5,
  databaseName = "sharable name of development data", outputDir)
}
\arguments{
\item{analysesDir}{The directory containing folders with PLP models}

\item{minCellCount}{The min cell count when trasporting the PLP model evaluation results}

\item{databaseName}{The name of the database as a string}

\item{outputDir}{the location to save the transported models (defaults to inst/plp_models)}
}
\value{
The models will now be in the package
}
\description{
Transport trained PLP models into the validation package
}
\details{
This will tranport PLP models into a validation package
}
